backend/requirements.txt
# Django core
Django>=5.1,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-filter>=24.0,<25.0
django-cors-headers>=4.4,<5.0
django-environ>=0.11,<1.0

# Database
psycopg2-binary>=2.9,<3.0

# Async / Tasks
celery>=5.4,<6.0
redis>=5.0,<6.0
django-celery-beat>=2.6,<3.0

# HTTP client (used by webhook emitter — Slice 22)
requests>=2.31,<3.0

# WebSocket
channels>=4.1,<5.0
daphne>=4.1,<5.0

# API docs
drf-spectacular>=0.27,<1.0

# Testing
pytest>=8.0,<9.0
pytest-django>=4.8,<5.0
pytest-cov>=5.0,<6.0
factory-boy>=3.3,<4.0

# Config / YAML
pyyaml>=6.0,<7.0

# Code quality
black>=24.0
isort>=5.13,<6.0
flake8>=7.0,<8.0
